Partially working solution to keeping IMAP connections alive.
However, the timer will only run once every 15 minutes on a random worker.
Since the connection pool is local to each worker, it needs to run
on each worker. The ideal solution would be to make the connection
pool global though this is probably not possible. Alternatively,
this resnet_internal.dailyduties.tasks.keep_imap_alive could  be made to
run on each worker rather than only once. This is the solution I'll be
investigating first.